Format dataframe to include correct case headers
Split case headers according to the pattern period and digit, the
string is split only if it ends with a period and digit after that.
This is used because pandas read duplicated columns by making them
unique. It does that by adding a period and digit after that.
splitting by only periods creae issues as periods might be in the
header itself. Also, .isin makes sure only desired case headers are
selected.